Introduce an applicative error type. Better type for emptySchema
Add tests for invalid input and fix bug in targeted decode

Do not consider further solutions to decode a recordfields.

This should never happen unless the source schema is ambiguous, as the target
schema tells us exactly what we need.

If we did not have a target schema, then we would need to consider all the
options.

Better handling of divergent operations

add dep on mtl

Add another test for recursive schemas